+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update AC500 support
 * Add split phase parsing support
 * Remove unsupported power_generation field from EB3A
+* Improved error reporting
 
 ## 0.9.2